refactor:여행생성로직 trip폴더로 이동
