+++ frontend/src/api/trip/postTrip.ts
@@ -3,7 +3,7 @@ import { NewTripData } from '@type/trips';
 
 import { axiosInstance } from '@api/axiosInstance';
 
-export const postNewTrip = () => async (newTripData: NewTripData) => {
+export const postTrip = () => async (newTripData: NewTripData) => {
   const response = await axiosInstance.post(END_POINTS.TRIPS, newTripData);
 
   return response.headers.location;
