+++ frontend/src/components/trip/TripCreateForm/TripCreateForm.style.ts
