+++ frontend/src/components/trip/TripCreateForm/TripCreateForm.tsx
@@ -1,4 +1,3 @@
-import { formStyling } from '@/components/newTrip/NewTripForm/TripCreateForm.style';
 import { PATH } from '@constants/path';
 import { Button } from 'hang-log-design-system';
 import type { FormEvent } from 'react';
@@ -9,6 +8,7 @@ import { useCityDateForm } from '@hooks/common/useCityDateForm';
 
 import CitySearchBar from '@components/common/CitySearchBar/CitySearchBar';
 import DateInput from '@components/common/DateInput/DateInput';
+import { formStyling } from '@components/trip/TripCreateForm/TripCreateForm.style';
 
 const TripCreateForm = () => {
   const { cityDateInfo, updateCityInfo, updateDateInfo, isCityDateValid } = useCityDateForm();
