+++ frontend/src/hooks/api/useCreateTripMutation.ts
@@ -1,10 +1,9 @@
+import { postTrip } from '@/api/trip/postTrip';
 import { ERROR_MESSAGE, NETWORK } from '@constants/api';
 import { useMutation } from '@tanstack/react-query';
 
-import { postNewTrip } from '@api/trips/postNewTrip';
-
 export const useCreateTripMutation = () => {
-  const newTripMutation = useMutation(postNewTrip(), {
+  const newTripMutation = useMutation(postTrip(), {
     onError: (err, _, context) => {
       //TODO:toast 띄우기
       alert(ERROR_MESSAGE);
